refactor: Move instrinics subdirectory under state-types
