+++ lib/state-machines/state-types/instrinsics/format.js
