+++ lib/state-machines/state-types/instrinsics/index.js
